market-bot/screen-monitor/requirements.txt
pillow
opencv-python-headless
mss
pytesseract
numpy
requests
